Refactor function flow to avoid having multiple actions
We previously were writing the transactions, parsing them and creating
the workbook on the same function, while having the
"compute_transactions" transactions name, which doesn't fit the bill for
all the actions that it was doing.

Separate the different actions into functions, that are easily testable
and can be extended.